@@ -14,62 +15,66 @@
     },
     "dependencies": {
         "@types/bcryptjs": "^2.4.2",
-        "@types/configstore": "^4.0.0",
+        "@types/configstore": "^5.0.1",
         "@types/cookie-parser": "^1.4.2",
-        "@types/dockerode": "^3.2.2",
-        "@types/ejs": "^3.0.5",
-        "@types/express": "^4.17.11",
-        "@types/fs-extra": "^9.0.7",
-        "@types/http-proxy": "^1.17.5",
-        "@types/jest": "^26.0.20",
-        "@types/js-base64": "^3.0.0",
-        "@types/jsonwebtoken": "^8.5.0",
-        "@types/morgan": "^1.9.2",
-        "@types/multer": "^1.4.5",
+        "@types/debug": "^4.1.7",
+        "@types/dockerode": "^3.2.7",
+        "@types/ejs": "^3.1.0",
+        "@types/express": "^4.17.13",
+        "@types/fs-extra": "^9.0.12",
+        "@types/http-proxy": "^1.17.7",
+        "@types/is-valid-path": "^0.1.0",
+        "@types/jest": "^27.0.1",
+        "@types/js-base64": "^3.3.1",
+        "@types/jsonwebtoken": "^8.5.5",
+        "@types/morgan": "^1.9.3",
+        "@types/multer": "^1.4.7",
         "@types/on-finished": "^2.3.1",
-        "@types/request": "^2.48.5",
-        "@types/require-from-string": "^1.2.0",
-        "@types/serve-favicon": "^2.5.2",
-        "@types/ssh2": "^0.5.46",
-        "@types/tar": "^4.0.4",
-        "@types/uuid": "^8.3.0",
-        "@types/validator": "^13.1.3",
+        "@types/request": "^2.48.7",
+        "@types/require-from-string": "^1.2.1",
+        "@types/serve-favicon": "^2.5.3",
+        "@types/ssh2": "^0.5.47",
+        "@types/tar": "^4.0.5",
+        "@types/uuid": "^8.3.1",
+        "@types/validator": "^13.6.3",
+        "@typescript-eslint/eslint-plugin": "^4.29.3",
+        "@typescript-eslint/parser": "^4.29.3",
         "axios": "^0.21.1",
         "bcryptjs": "^2.4.3",
         "body-parser": "^1.19.0",
         "configstore": "^5.0.1",
         "cookie-parser": "~1.4.5",
-        "debug": "~4.3.1",
-        "dockerode": "^3.2.1",
+        "debug": "~4.3.2",
+        "dockerode": "^3.3.0",
         "ejs": "^3.1.6",
         "express": "^4.17.1",
-        "fs-extra": "^9.1.0",
+        "fs-extra": "^10.0.0",
         "http-proxy": "^1.18.1",
         "is-valid-path": "^0.1.1",
-        "js-base64": "^3.6.0",
+        "js-base64": "^3.6.1",
         "jsonwebtoken": "^8.5.1",
-        "madge": "^4.0.0",
+        "madge": "^5.0.1",
         "moment": "^2.29.1",
         "morgan": "^1.10.0",
-        "multer": "^1.4.2",
+        "multer": "^1.4.3",
         "on-finished": "^2.3.0",
-        "prettier": "^2.2.1",
-        "public-ip": "^4.0.3",
+        "prettier": "2.3.2",
+        "public-ip": "^4.0.4",
         "recursive-readdir": "^2.2.2",
         "request": "^2.88.2",
         "require-from-string": "^2.0.2",
         "serve-favicon": "~2.5.0",
-        "simple-git": "^2.34.2",
-        "ssh2": "^0.8.9",
-        "tar": "^6.1.0",
-        "tslint": "^6.1.3",
-        "typescript": "^4.1.5",
+        "simple-git": "^2.45.0",
+        "ssh2": "^1.3.0",
+        "tar": "^6.1.11",
+        "typescript": "^4.3.5",
         "uuid": "^8.3.2",
-        "validator": "^13.5.2",
-        "yaml": "^1.10.0"
+        "validator": "^13.6.0",
+        "yaml": "^1.10.2"
     },
     "devDependencies": {
-        "jest": "^26.6.3",
-        "ts-jest": "^26.5.1"
+        "eslint": "^7.32.0",
+        "jest": "^27.1.0",
+        "ts-jest": "^27.0.5"
     }
 }
