@@ -18,16 +18,16 @@ const DEFAULT_APP_CAPTAIN_DEFINITION = JSON.stringify({
 
 // unused images
 router.get('/unusedImages', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
     Promise.resolve()
         .then(function () {
-            let mostRecentLimit = Number(req.query.mostRecentLimit || '0')
+            const mostRecentLimit = Number(req.query.mostRecentLimit || '0')
             return serviceManager.getUnusedImages(mostRecentLimit)
         })
         .then(function (unusedImages) {
-            let baseApi = new BaseApi(
+            const baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Unused images retrieved.'
             )
@@ -41,16 +41,16 @@ router.get('/unusedImages', function (req, res, next) {
 
 // delete images
 router.post('/deleteImages', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
-    let imageIds = req.body.imageIds || []
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
+    const imageIds = req.body.imageIds || []
 
     Promise.resolve()
         .then(function () {
             return serviceManager.deleteImages(imageIds)
         })
         .then(function () {
-            let baseApi = new BaseApi(
+            const baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'Images Deleted.'
             )
@@ -61,20 +61,20 @@ router.post('/deleteImages', function (req, res, next) {
 
 // Get All App Definitions
 router.get('/', function (req, res, next) {
-    const dataStore = InjectionExtractor.extractUserFromInjected(res).user
-        .dataStore
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
-    let appsArray: IAppDef[] = []
+    const dataStore =
+        InjectionExtractor.extractUserFromInjected(res).user.dataStore
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
+    const appsArray: IAppDef[] = []
 
     dataStore
         .getAppsDataStore()
         .getAppDefinitions()
         .then(function (apps) {
-            let promises: Promise<void>[] = []
+            const promises: Promise<void>[] = []
 
             Object.keys(apps).forEach(function (key, index) {
-                let app = apps[key]
+                const app = apps[key]
                 app.appName = key
                 app.isAppBuilding = serviceManager.isAppBuilding(key)
                 app.appPushWebhook = app.appPushWebhook || undefined
@@ -87,7 +87,7 @@ router.get('/', function (req, res, next) {
             return dataStore.getDefaultAppNginxConfig()
         })
         .then(function (defaultNginxConfig) {
-            let baseApi = new BaseApi(
+            const baseApi = new BaseApi(
                 ApiStatusCodes.STATUS_OK,
                 'App definitions are retrieved.'
             )
@@ -103,8 +103,8 @@ router.get('/', function (req, res, next) {
 })
 
 router.post('/enablebasedomainssl/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
     const appName = req.body.appName
 
@@ -113,7 +113,7 @@ router.post('/enablebasedomainssl/', function (req, res, next) {
             return serviceManager.enableSslForApp(appName)
         })
         .then(function () {
-            let msg = `General SSL is enabled for: ${appName}`
+            const msg = `General SSL is enabled for: ${appName}`
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
         })
@@ -121,11 +121,11 @@ router.post('/enablebasedomainssl/', function (req, res, next) {
 })
 
 router.post('/customdomain/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
-    let appName = req.body.appName
-    let customDomain = (req.body.customDomain || '').toLowerCase()
+    const appName = req.body.appName
+    const customDomain = (req.body.customDomain || '').toLowerCase()
 
     // verify customdomain.com going through the default NGINX
     // Add customdomain.com to app in Data Store
@@ -135,7 +135,7 @@ router.post('/customdomain/', function (req, res, next) {
             return serviceManager.addCustomDomain(appName, customDomain)
         })
         .then(function () {
-            let msg = `Custom domain is enabled for: ${appName} at ${customDomain}`
+            const msg = `Custom domain is enabled for: ${appName} at ${customDomain}`
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
         })
@@ -143,18 +143,18 @@ router.post('/customdomain/', function (req, res, next) {
 })
 
 router.post('/removecustomdomain/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
-    let appName = req.body.appName
-    let customDomain = (req.body.customDomain || '').toLowerCase()
+    const appName = req.body.appName
+    const customDomain = (req.body.customDomain || '').toLowerCase()
 
     return Promise.resolve()
         .then(function () {
             return serviceManager.removeCustomDomain(appName, customDomain)
         })
         .then(function () {
-            let msg = `Custom domain is removed for: ${appName} at ${customDomain}`
+            const msg = `Custom domain is removed for: ${appName} at ${customDomain}`
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
         })
@@ -162,11 +162,11 @@ router.post('/removecustomdomain/', function (req, res, next) {
 })
 
 router.post('/enablecustomdomainssl/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
-    let appName = req.body.appName
-    let customDomain = (req.body.customDomain || '').toLowerCase()
+    const appName = req.body.appName
+    const customDomain = (req.body.customDomain || '').toLowerCase()
 
     // Check if customdomain is already associated with app. If not, error out.
     // Verify customdomain.com is served from /customdomain.com/
@@ -176,7 +176,7 @@ router.post('/enablecustomdomainssl/', function (req, res, next) {
             return serviceManager.enableCustomDomainSsl(appName, customDomain)
         })
         .then(function () {
-            let msg = `Custom domain SSL is enabled for: ${appName} at ${customDomain} `
+            const msg = `Custom domain SSL is enabled for: ${appName} at ${customDomain} `
             Logger.d(msg)
             res.send(new BaseApi(ApiStatusCodes.STATUS_OK, msg))
         })
@@ -184,10 +184,10 @@ router.post('/enablecustomdomainssl/', function (req, res, next) {
 })
 
 router.post('/register/', function (req, res, next) {
-    const dataStore = InjectionExtractor.extractUserFromInjected(res).user
-        .dataStore
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const dataStore =
+        InjectionExtractor.extractUserFromInjected(res).user.dataStore
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
     const appName = req.body.appName as string
     const hasPersistentData = !!req.body.hasPersistentData
@@ -208,7 +208,8 @@ router.post('/register/', function (req, res, next) {
                 appName,
                 {
                     captainDefinitionContentSource: {
-                        captainDefinitionContent: DEFAULT_APP_CAPTAIN_DEFINITION,
+                        captainDefinitionContent:
+                            DEFAULT_APP_CAPTAIN_DEFINITION,
                         gitHash: '',
                     },
                 }
@@ -244,11 +245,11 @@ router.post('/register/', function (req, res, next) {
 })
 
 router.post('/delete/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
-    let appName = req.body.appName
-    let volumes = req.body.volumes || []
+    const appName = req.body.appName
+    const volumes = req.body.volumes || []
 
     Logger.d(`Deleting app started: ${appName}`)
 
@@ -284,11 +285,11 @@ router.post('/delete/', function (req, res, next) {
 })
 
 router.post('/rename/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
-    let oldAppName = req.body.oldAppName + ''
-    let newAppName = req.body.newAppName + ''
+    const oldAppName = req.body.oldAppName + ''
+    const newAppName = req.body.newAppName + ''
 
     Logger.d(`Renaming app started: From ${oldAppName} To ${newAppName} `)
 
@@ -306,29 +307,45 @@ router.post('/rename/', function (req, res, next) {
 })
 
 router.post('/update/', function (req, res, next) {
-    const serviceManager = InjectionExtractor.extractUserFromInjected(res).user
-        .serviceManager
+    const serviceManager =
+        InjectionExtractor.extractUserFromInjected(res).user.serviceManager
 
-    let appName = req.body.appName
-    let nodeId = req.body.nodeId
-    let captainDefinitionRelativeFilePath =
+    const appName = req.body.appName
+    const nodeId = req.body.nodeId
+    const captainDefinitionRelativeFilePath =
         req.body.captainDefinitionRelativeFilePath
-    let notExposeAsWebApp = req.body.notExposeAsWebApp
-    let customNginxConfig = req.body.customNginxConfig
-    let forceSsl = !!req.body.forceSsl
-    let websocketSupport = !!req.body.websocketSupport
-    let repoInfo = !!req.body.appPushWebhook
+    const notExposeAsWebApp = req.body.notExposeAsWebApp
+    const customNginxConfig = req.body.customNginxConfig
+    const forceSsl = !!req.body.forceSsl
+    const websocketSupport = !!req.body.websocketSupport
+    const repoInfo = req.body.appPushWebhook
         ? req.body.appPushWebhook.repoInfo || {}
         : {}
-    let envVars = req.body.envVars || []
-    let volumes = req.body.volumes || []
-    let ports = req.body.ports || []
-    let instanceCount = req.body.instanceCount || '0'
-    let preDeployFunction = req.body.preDeployFunction || ''
-    let serviceUpdateOverride = req.body.serviceUpdateOverride || ''
-    let containerHttpPort = Number(req.body.containerHttpPort) || 80
-    let httpAuth = req.body.httpAuth
-    let description = req.body.description || ''
+    const envVars = req.body.envVars || []
+    const volumes = req.body.volumes || []
+    const ports = req.body.ports || []
+    const instanceCount = req.body.instanceCount || '0'
+    const preDeployFunction = req.body.preDeployFunction || ''
+    const serviceUpdateOverride = req.body.serviceUpdateOverride || ''
+    const containerHttpPort = Number(req.body.containerHttpPort) || 80
+    const httpAuth = req.body.httpAuth
+    let appDeployTokenConfig = req.body.appDeployTokenConfig as
+        | AppDeployTokenConfig
+        | undefined
+    const description = req.body.description || ''
+
+    if (!appDeployTokenConfig) {
+        appDeployTokenConfig = { enabled: false }
+    } else {
+        appDeployTokenConfig = {
+            enabled: !!appDeployTokenConfig.enabled,
+            appDeployToken: `${
+                appDeployTokenConfig.appDeployToken
+                    ? appDeployTokenConfig.appDeployToken
+                    : ''
+            }`.trim(),
+        }
+    }
 
     if (repoInfo.user) {
         repoInfo.user = repoInfo.user.trim()
@@ -381,7 +398,8 @@ router.post('/update/', function (req, res, next) {
             customNginxConfig,
             preDeployFunction,
             serviceUpdateOverride,
-            websocketSupport
+            websocketSupport,
+            appDeployTokenConfig
         )
         .then(function () {
             Logger.d(`AppName is updated: ${appName}`)
