@@ -42,7 +42,7 @@ class DockerRegistryHelper {
                 return self.getDefaultPushRegistryId()
             })
             .then(function (defaultRegId) {
-                let ret: IRegistryInfo | undefined = undefined
+                const ret: IRegistryInfo | undefined = undefined
                 for (let idx = 0; idx < allRegistries.length; idx++) {
                     const element = allRegistries[idx]
                     if (defaultRegId && element.id === defaultRegId) {
@@ -133,18 +133,26 @@ class DockerRegistryHelper {
                     }
                 }
 
-                // if none of the registries explicitly relates to the image name, try the official one
-                for (let index = 0; index < regs.length; index++) {
-                    const element = regs[index]
-                    if (element.registryDomain === 'registry-1.docker.io') {
-                        return {
-                            serveraddress: element.registryDomain,
-                            username: element.registryUser,
-                            password: element.registryPassword,
-                            // email: CaptainConstants.defaultEmail, // email is optional
+                // if none of the registries explicitly relates to the image name, and no other explicit domain is defined,
+                // try Docker Hub registry as the default
+                if (
+                    imageName.split('/').length == 2 || // user/image is from Docker Hub
+                    imageName.split('/')[0].endsWith('.docker.io') || // registry-1.docker.io/user/image is from Docker Hub
+                    imageName.split('/')[0] === 'docker.io' || // registry-1.docker.io/user/image is from Docker Hub
+                    imageName.split('/')[0].endsWith('.docker.com') || // hub.docker.com/user/image is from Docker Hub
+                    imageName.split('/')[0] === 'docker.io' // hub.docker.com/user/image is from Docker Hub
+                )
+                    for (let index = 0; index < regs.length; index++) {
+                        const element = regs[index]
+                        if (element.registryDomain === 'registry-1.docker.io') {
+                            return {
+                                serveraddress: element.registryDomain,
+                                username: element.registryUser,
+                                password: element.registryPassword,
+                                // email: CaptainConstants.defaultEmail, // email is optional
+                            }
                         }
                     }
-                }
 
                 return undefined
             })
@@ -158,7 +166,7 @@ class DockerRegistryHelper {
                 return self.getAllRegistries()
             })
             .then(function (regs) {
-                let registryConfig: DockerRegistryConfig = {}
+                const registryConfig: DockerRegistryConfig = {}
 
                 for (let index = 0; index < regs.length; index++) {
                     const element = regs[index]
@@ -254,32 +262,24 @@ class DockerRegistryHelper {
                     return
                 }
 
-                return self.dockerApi
-                    .checkRegistryAuth({
-                        username: registryUser,
-                        password: registryPassword,
-                        serveraddress: registryDomain,
-                        // email: CaptainConstants.defaultEmail, // email is optional
-                    })
-                    .catch(function (err) {
-                        Logger.e(err)
-                        throw ApiStatusCodes.createError(
-                            ApiStatusCodes.AUTHENTICATION_FAILED,
-                            'Authentication failed. Either username, password or domain is incorrect.'
-                        )
-                    })
+                return self.ensureAuthenticationForRegistry(
+                    registryUser,
+                    registryPassword,
+                    registryDomain
+                )
             })
             .then(function () {
                 return self.registriesDataStore.getAllRegistries()
             })
             .then(function (allRegs) {
-                let promiseToAddRegistry = self.registriesDataStore.addRegistryToDb(
-                    registryUser,
-                    registryPassword,
-                    registryDomain,
-                    registryImagePrefix,
-                    registryType
-                )
+                let promiseToAddRegistry =
+                    self.registriesDataStore.addRegistryToDb(
+                        registryUser,
+                        registryPassword,
+                        registryDomain,
+                        registryImagePrefix,
+                        registryType
+                    )
 
                 // Product decision. We want to make the first added registry the default one,
                 // this way, it's easier for new users to grasp the concept of default push registry.
@@ -306,17 +306,46 @@ class DockerRegistryHelper {
         registryImagePrefix: string
     ) {
         const self = this
-        return Promise.resolve().then(function () {
-            registryDomain = Utils.removeHttpHttps(registryDomain)
+        return Promise.resolve()
+            .then(function () {
+                registryDomain = Utils.removeHttpHttps(registryDomain)
 
-            return self.registriesDataStore.updateRegistry(
-                id,
-                registryUser,
-                registryPassword,
-                registryDomain,
-                registryImagePrefix
-            )
-        })
+                return self.ensureAuthenticationForRegistry(
+                    registryUser,
+                    registryPassword,
+                    registryDomain
+                )
+            })
+            .then(function () {
+                return self.registriesDataStore.updateRegistry(
+                    id,
+                    registryUser,
+                    registryPassword,
+                    registryDomain,
+                    registryImagePrefix
+                )
+            })
+    }
+
+    private ensureAuthenticationForRegistry(
+        registryUser: string,
+        registryPassword: string,
+        registryDomain: string
+    ): any {
+        const self = this
+        return self.dockerApi
+            .checkRegistryAuth({
+                username: registryUser,
+                password: registryPassword,
+                serveraddress: registryDomain,
+            })
+            .catch(function (err) {
+                Logger.e(err)
+                throw ApiStatusCodes.createError(
+                    ApiStatusCodes.AUTHENTICATION_FAILED,
+                    'Authentication failed. Either username, password or domain is incorrect.'
+                )
+            })
     }
 }
 
