@@ -205,7 +205,7 @@ class CaptainManager {
                 )
                     .migrateIfNeeded()
                     .then(function (migrationPerformed) {
-                        if (!!migrationPerformed) {
+                        if (migrationPerformed) {
                             return self.resetSelf()
                         }
                     })
@@ -226,7 +226,7 @@ class CaptainManager {
                     }
                 }
 
-                if (!!localRegistry) {
+                if (localRegistry) {
                     Logger.d('Ensuring Docker Registry is running...')
                     return self.dockerRegistry.ensureDockerRegistryRunningOnThisNode(
                         localRegistry.registryPassword
@@ -268,7 +268,7 @@ class CaptainManager {
     performHealthCheck() {
         const self = this
         const captainPublicDomain = `${
-            CaptainConstants.captainSubDomain
+            CaptainConstants.configs.captainSubDomain
         }.${self.dataStore.getRootDomain()}`
 
         function scheduleNextHealthCheck() {
@@ -660,7 +660,7 @@ class CaptainManager {
             .then(function () {
                 return self.certbotManager.enableSsl(
                     `${
-                        CaptainConstants.captainSubDomain
+                        CaptainConstants.configs.captainSubDomain
                     }.${self.dataStore.getRootDomain()}`
                 )
             })
@@ -786,7 +786,7 @@ class CaptainManager {
                     }
                 }
 
-                if (!!localRegistry) {
+                if (localRegistry) {
                     throw ApiStatusCodes.createError(
                         ApiStatusCodes.ILLEGAL_OPERATION,
                         'Delete your self-hosted Docker registry before changing the domain.'
