@@ -4,7 +4,7 @@ test('Testing Encryptor 1', () => {
     const encryptor = new CaptainEncryptor(
         '8h9hasfasaaaaaaaaaaaaaa75h7553245235423452345235235235254h75h38'
     )
-    let valueToBeEncrypter = 'qq'
+    const valueToBeEncrypter = 'qq'
     expect(encryptor.decrypt(encryptor.encrypt(valueToBeEncrypter))).toBe(
         valueToBeEncrypter
     )
@@ -14,7 +14,7 @@ test('Testing Encryptor 2', () => {
     const encryptor = new CaptainEncryptor(
         '8h9hasfasaaaaaaaaaaaaaa75h7553245235423452345235235235254h75h38'
     )
-    let valueToBeEncrypter = 'q290852f98nb80nv8m8m bn83vn@ 8098m%#@%$5$@#52q'
+    const valueToBeEncrypter = 'q290852f98nb80nv8m8m bn83vn@ 8098m%#@%$5$@#52q'
     expect(encryptor.decrypt(encryptor.encrypt(valueToBeEncrypter))).toBe(
         valueToBeEncrypter
     )
